requirements.txt
PyHamcrest==2.0.2
allure-pytest==2.8.11
allure-python-commons==2.8.11
allure-python-commons-test==2.8.11
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
idna==2.9
importlib-metadata==1.5.0
joblib==0.14.1
kiwisolver==1.1.0
matplotlib==3.2.0
more-itertools==8.2.0
numpy==1.18.1
packaging==20.3
pandas==1.0.2
pip==19.0.3
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.6
pytest==5.4.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
scikit-learn==0.22.2.post1
scipy==1.4.1
selenium==3.141.0
setuptools==40.8.0
six==1.14.0
urllib3==1.25.8
wcwidth==0.1.8
xlwt==1.3.0
zipp==3.1.0